+++ vcxproj2cmake.Tests/ConverterTests/EnableStandaloneProjectBuildsTests.cs
@@ -37,7 +37,7 @@ public partial class ConverterTests
                 add_executable(App)
 
                 target_link_libraries(App
-                    PUBLIC
+                    PRIVATE
                         Lib
                 )
                 """);