+++ vcxproj2cmake.Tests/ConverterTests/RemoveObsoleteLibrariesFromProjectReferencesTests.cs
@@ -65,7 +65,7 @@ public partial class ConverterTests
                 add_executable(App)
 
                 target_link_libraries(App
-                    PUBLIC
+                    PRIVATE
                         Lib.lib
                 )
                 """);
@@ -96,7 +96,7 @@ public partial class ConverterTests
                 add_executable(App)
 
                 target_link_libraries(App
-                    PUBLIC
+                    PRIVATE
                         Lib
                 )
                 """);
@@ -131,7 +131,7 @@ public partial class ConverterTests
                 add_executable(App)
 
                 target_link_libraries(App
-                    PUBLIC
+                    PRIVATE
                         Lib
                 )
                 """);