+++ vcxproj2cmake.Tests/ConverterTests/ResolveProjectReferencesTests.cs
@@ -57,7 +57,7 @@ public partial class ConverterTests
                 add_executable(App)
 
                 target_link_libraries(App
-                    PUBLIC
+                    PRIVATE
                         Lib
                 )
                 """);
@@ -95,7 +95,7 @@ public partial class ConverterTests
                 add_executable(App)
 
                 target_link_libraries(App
-                    PUBLIC
+                    PRIVATE
                         Lib
                 )
                 """);